+++ Ejemplo19.java
@@ -2,15 +2,15 @@ import java.util.Scanner;
 public class Ejemplo19 {
     public static void main(String[] args) {
         
-        int precioH, horas, pago_bruto, pago_neto, impuestos;
+        double precioH, horas, pago_bruto, pago_neto, impuestos;
 
         System.out.println("Dame el precio/h");
         Scanner inputValue;
         inputValue = new Scanner(System.in);
-        precioH = inputValue.nextInt();
+        precioH = inputValue.nextDouble();
         
         System.out.println("Dame el numero");
-        horas = inputValue.nextInt();
+        horas = inputValue.nextDouble();
 
         
     }
